+++ pkg/pool-stable/contracts/ComposableStablePoolProtocolFees.sol
@@ -317,7 +317,7 @@ abstract contract ComposableStablePoolProtocolFees is
         override(
             // Our inheritance pattern creates a small diamond that requires explicitly listing the parents here.
             // Each parent calls the `super` version, so linearization ensures all implementations are called.
-            LegacyBasePool,
+            BasePool,
             BasePoolAuthorization,
             ComposableStablePoolRates
         )
