+++ pkg/pool-stable/contracts/ComposableStablePoolStorage.sol
@@ -18,11 +18,11 @@ import "@balancer-labs/v2-interfaces/contracts/solidity-utils/helpers/BalancerEr
 import "@balancer-labs/v2-interfaces/contracts/solidity-utils/openzeppelin/IERC20.sol";
 import "@balancer-labs/v2-interfaces/contracts/pool-utils/IRateProvider.sol";
 
-import "@balancer-labs/v2-pool-utils/contracts/LegacyBasePool.sol";
+import "@balancer-labs/v2-pool-utils/contracts/BasePool.sol";
 
 import "./StableMath.sol";
 
-abstract contract ComposableStablePoolStorage is LegacyBasePool {
+abstract contract ComposableStablePoolStorage is BasePool {
     using FixedPoint for uint256;
     using WordCodec for bytes32;
