+++ pkg/pool-stable/contracts/test/MockComposableStablePoolRates.sol
@@ -31,14 +31,14 @@ contract MockComposableStablePoolRates is ComposableStablePoolRates {
         ComposableStablePoolStorage(
             StorageParams(_insertSorted(tokens, IERC20(this)), tokenRateProviders, exemptFromYieldProtocolFeeFlags)
         )
-        LegacyBasePool(
+        BasePool(
             vault,
             IVault.PoolSpecialization.GENERAL,
             "MockStablePoolStorage",
             "MOCK_BPT",
             _insertSorted(tokens, IERC20(this)),
             new address[](tokens.length + 1),
-            1e12, // LegacyBasePool._MIN_SWAP_FEE_PERCENTAGE
+            1e12, // BasePool._MIN_SWAP_FEE_PERCENTAGE
             0,
             0,
             owner
