+++ pkg/pool-stable/contracts/test/MockComposableStablePoolStorage.sol
@@ -33,14 +33,14 @@ contract MockComposableStablePoolStorage is ComposableStablePoolStorage {
                 exemptFromYieldProtocolFeeFlags: exemptFromYieldProtocolFeeFlags
             })
         )
-        LegacyBasePool(
+        BasePool(
             vault,
             IVault.PoolSpecialization.GENERAL,
             "MockComposableStablePoolStorage",
             "MOCK_BPT",
             _insertSorted(tokens, IERC20(this)),
             new address[](tokens.length + 1),
-            1e12, // LegacyBasePool._MIN_SWAP_FEE_PERCENTAGE
+            1e12, // BasePool._MIN_SWAP_FEE_PERCENTAGE
             0,
             0,
             address(0)
